Update package dependencies and engine versions
Updated all package dependencies to their latest versions in `package.json`. Also, improved the engines version to be more precise and compatible with `node` and `npm` versions. This is to make sure the current codebase is up to date, hence reducing potential risks and issues that could arise due to outdated packages or incompatible engine versions.